Dedupe core response setup
