+++ src/machines/timer_state_machine.rs
@@ -343,7 +343,6 @@ mod test {
         let commands = t
             .handle_workflow_task_take_cmds(&mut state_machines, Some(2))
             .unwrap();
-        dbg!(&commands);
         assert_eq!(commands.len(), 1);
         assert_eq!(
             commands[0].command_type,
